:ant: [examples] Fix compile error
