+++ .gitignore
@@ -24,3 +24,4 @@ _testmain.go
 *.prof
 
 bin/
+.DS_Store
